This is to fix PAXLOGGING-142 where the PaxContext.get() method returns a String instance instead of an Object.  The original MDC class from log4j returns an Object but this version does not.
It seems like the change was done because the slf4j MDCAdapter class get method returns
a String and not an Object.

That caused the issue that the user will not be able to assign an Object instance for
a given key using the MDC class.

This fix updates the PaxContext get method to return an Object instead of a String.

The Slf4jMDCAdapter get method still returns a String but now checks if the return
instance from MDC get method is non-null and then calls its toString method.

This change keeps things consistent throughout and allows other to add a
non-String instance using MDC.

This does not mean that the Slf4jMDCAdapter put can now set an Object as the value.

It still requires a String value.