This fixes PAXLOGGING-142, where the PaxContext get method does not return an Object but the PaxContext put method allows an Object to be passed as a value.
The get method now returns an Object instance and the Slf4jMDCAdapter get method
was changed to handle that case.

The MDCTest test was added to check that this change does not break the
interface.

Users can now use the MDC class to put an Object and retrieve that Object
instance using the MDC get method.